refactor: split MySQL metadata responsibilities
